@@ -4,7 +4,7 @@ version: 0.0.1
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.18.6 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
