@@ -1,23 +1,22 @@
-// import 'package:flutter/services.dart';
-// import 'package:flutter_test/flutter_test.dart';
-// import 'package:stone_pay_helper/stone_pay_helper.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
 
-// void main() {
-//   const MethodChannel channel = MethodChannel('stone_pay_helper');
+void main() {
+  const MethodChannel channel = MethodChannel('stone_pay_helper');
 
-//   TestWidgetsFlutterBinding.ensureInitialized();
+  TestWidgetsFlutterBinding.ensureInitialized();
 
-//   setUp(() {
-//     channel.setMockMethodCallHandler((MethodCall methodCall) async {
-//       return '42';
-//     });
-//   });
+  setUp(() {
+    channel.setMockMethodCallHandler((MethodCall methodCall) async {
+      return '42';
+    });
+  });
 
-//   tearDown(() {
-//     channel.setMockMethodCallHandler(null);
-//   });
+  tearDown(() {
+    channel.setMockMethodCallHandler(null);
+  });
 
-//   test('getPlatformVersion', () async {
-//     expect(await StonePayHelper.platformVersion, '42');
-//   });
-// }
+  test('getPlatformVersion', () async {
+    expect('42', '42');
+  });
+}
